requirements.txt
# local package
-e .

# external requirements
click
Sphinx
sphinx-rtd-theme
coverage
flake8
python-dotenv>=0.5.1
matplotlib
requests
scikit-learn
numpy
pandas
tqdm
azureml-sdk
azure-mgmt-resource
azure-mgmt-datalake-store
azure-datalake-store
azure-keyvault-secrets
torch
torchvision
mlflow